Completed the WordleFX class. The wordClean, checkWord, uniqueCharacters
and clearBoard method have all been implemented. This has yet to be

thoroughly tested with the actual user interface and that is the

next step that must be tackled, setting up the actual client code

to interact with the WordleFX method. Edits on the square class are

edits to the field values and images.