Moved the two data files to the proper position in the repository
under the actual data folder rather than one that I created. I fixed the

LoginSystem algorithms so that they now work and both the JavaFX and

the actual logic work for the login information as well as the files.

the NYTimesFX now has multiple scenes as well, the home page has been

created. The next step is to add the WordleFX and the shop, after which

the design is needed to be done.