Made significant progress on the Wordle board. Created the board and
managed to bring it to a point where letters could not be easily

outputted onto the board. Began connecting the client code with the

WordleFX code which is the last "major" hurdle of this project in terms

of pure thought process; however, in terms of time I believe that

stylizing the program and ensuring overall functionality will also

be another challenging task to be faced.