Finished stylizing the login pages and most of the homepage; however
since the rewards shop has yet to be implemented the home page has

also yet to be completely stylized as well. Some progress was made on

connecting the item shop redemption to the NYTimesFX client code. This
will most likely be the last step of the actual thought process behind
the programming and the remaining of time spent programming will put to
input validation, more stylization and appropriate documentation.