Figured out how to enter ENUMS into files. Resolved an issue in which
the binary search would not correctly find a username (this was the

result of the fact that the array was not sorted prior to the search),

a method was called to fix this issue; however, provided the time this

will be changed to an actual search algorithm. Furthermore, I toyed

around with implementing the backspace feature to the Wordle portion of

the program; however, this was unscuessful.