Did not finish comment. For some reason, no matter what is done the
file does not detect the word properly; however, this is only true

some of the time for some reason. I even printed the entire list into

the console which showed that the word was in fact in the list; yet the

program for some reason or another refused to believe this. Also during

this commit I finished all the design aspects which explains all the
commits with images.